t-target-api: Fix testcase when there is only 1 or 2 devices (#1)
The code sets the default-device-var ICV to 3 – assuming that host
fallback is done; well, that's the case for the 'target' regions but
not for 'omp_target_alloc' - as that requires a conforming device
number.

The solution is to call omp_target_alloc for the host if default-device-var
is too large, i.e. instead of checking for no device, the check is now
whether the default device is larger than the number of devices.

While that fixes the issue, I guess it is more useful to actually use
a device - if there is one or two available - which is the second change.

To check on the host, one way out is still to set
OMP_TARGET_OFFLOAD=DISABLED, which is the supposed way in omptests;
cf. utilities/check.h